Started operating with abstract class to extend
