+++ src/view/no-film-card.js
@@ -1,23 +1,10 @@
-import { createElement } from '../render';
+import AbstractView from "./abstract";
 
 const createNoFilmCardTemplate = () => '<h2 class="films-list__title">There are no movies in our database</h2>';
 
-export default class NoFilmCard {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class NoFilmCard extends AbstractView {
   get template() {
     return createNoFilmCardTemplate();
   }
-
-  removeElement() {
-    this.#element = null;
-  }
 }
+
