+++ src/view/show-more-btn.js
@@ -1,28 +1,14 @@
-import { createElement } from '../render';
+import AbstractView from "./abstract";
 
 const createShowMoreBtnTemplate = () => (
   `<button class='films-list__show-more'>Show more</button>
   `
 );
-
-export default class ShowMoreBtn {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class ShowMoreBtn extends AbstractView {
   get template() {
     return createShowMoreBtnTemplate();
   }
-
-  removeElement() {
-    this.#element = null;
-  }
 }
+
